// ui/src/openapi/models/ApiTaskStats.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ShpanKids API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ApiTaskStats
 */
export interface ApiTaskStats {
    /**
     * 
     * @type {Date}
     * @memberof ApiTaskStats
     */
    forDate: Date;
    /**
     * 
     * @type {string}
     * @memberof ApiTaskStats
     */
    userId: string;
    /**
     * 
     * @type {number}
     * @memberof ApiTaskStats
     */
    doneTasksCount: number;
    /**
     * 
     * @type {number}
     * @memberof ApiTaskStats
     */
    totalTasksCount: number;
}

/**
 * Check if a given object implements the ApiTaskStats interface.
 */
export function instanceOfApiTaskStats(value: object): boolean {
    if (!('forDate' in value)) return false;
    if (!('userId' in value)) return false;
    if (!('doneTasksCount' in value)) return false;
    if (!('totalTasksCount' in value)) return false;
    return true;
}

export function ApiTaskStatsFromJSON(json: any): ApiTaskStats {
    return ApiTaskStatsFromJSONTyped(json, false);
}

export function ApiTaskStatsFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiTaskStats {
    if (json == null) {
        return json;
    }
    return {
        
        'forDate': (new Date(json['forDate'])),
        'userId': json['userId'],
        'doneTasksCount': json['doneTasksCount'],
        'totalTasksCount': json['totalTasksCount'],
    };
}

export function ApiTaskStatsToJSON(value?: ApiTaskStats | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'forDate': ((value['forDate']).toISOString()),
        'userId': value['userId'],
        'doneTasksCount': value['doneTasksCount'],
        'totalTasksCount': value['totalTasksCount'],
    };
}